+++ caffe2/CMakeLists.txt
@@ -29,9 +29,6 @@ add_subdirectory(sgd)
 add_subdirectory(share)
 # add_subdirectory(test) # todo: use caffe2_gtest_main instead of gtest_main because we will need to call GlobalInit
 add_subdirectory(transforms)
-if (USE_TENSORRT)
-add_subdirectory(trt)
-endif()
 add_subdirectory(utils)
 
 # Advanced: if we have white list specified, we will do intersections for all
