+++ cmake/Summary.cmake
@@ -50,7 +50,6 @@ function (caffe2_print_configuration_summary)
     if(${USE_TENSORRT})
       message(STATUS "      TensorRT runtime library: ${TENSORRT_LIBRARY}")
       message(STATUS "      TensorRT include path   : ${TENSORRT_INCLUDE_DIR}")
-      message(STATUS "      ONNX_TRT library: ${ONNX_TRT_LIBRARY}")
     endif()
   endif()
   message(STATUS "  USE_EIGEN_FOR_BLAS    : ${CAFFE2_USE_EIGEN_FOR_BLAS}")
